Move search configuration into a separate file
- This also switches to the `experimental-typescript` tag of `instansearch.js`
